mac80211: mesh power save doze scheduling
Configure the HW for PS mode if the local mesh PS parameters allow it.

Expose a callback mps_hw_init for drivers to register mesh powersave ops:
hw_doze - put the radio to sleep now
hw_wakeup - wake the radio up for frame RX
These may be extended in the future to allow drivers/HW to implement mesh PS
themselves. The current design goal was to concentrate mesh PS routines in
mac80211 to keep driver modifications minimal.

Track the timing information of beacons from peers we are in PS mode towards.
Set a per-STA timer which will trigger a wakeup right before the peer's next
TBTT.
Also use the timer to go to sleep mode after not receiving a beacon in a
defined time margin. In this case project the next TBTT and increase the
margin.

For mesh Awake Window start a timer on SWBA (beacon_get_tim) which on
expiration triggers a hw_doze call.

Expose a callback mps_hw_doze allow, which allows the driver to request
doze state in mesh mode. This is used to keep the HW awake when sending
frames.
Since this function may be called very frequently, maintain status variables
in the interface struct to avoid iterating the STA list everytime.

Signed-off-by: Marco Porsch <marco@cozybit.com>